service/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
line-bot-sdk==2.4.2
multidict==6.0.4
Pillow==9.5.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.9.1
